[loading] Split directory loader into multiple files
