+++ src/loading/dir/mod.rs
@@ -1,5 +1,5 @@
 use crate::core::reactor::SingleThreadReactor;
-use crate::fs;
+use crate::fs::dir::{DirectoryEntry, DirectoryTree};
 use crate::loading::{FileTree, LoadingError};
 use futures::Future;
 use matches::matches;
@@ -7,40 +7,20 @@ use std::collections::hash_map;
 use std::collections::HashMap;
 use std::ffi::OsString;
 use std::io;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::sync::Arc;
+
+mod iter;
+mod reactor;
+
+pub use iter::*;
+use reactor::*;
 
 pub struct DirectoryFileTree(Arc<DirectoryFileTreeData>);
 
 struct DirectoryFileTreeData {
-    cache: DirectoryCache,
+    cache: DirectoryTree,
     reactor: SingleThreadReactor<FileSystemOp, FileSystemOpResult>,
-}
-
-struct DirectoryCache {
-    root: PathBuf,
-    entry: DirectoryEntry,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum DirectoryEntry {
-    Directory(HashMap<OsString, DirectoryEntry>),
-    File,
-}
-
-enum FileSystemOp {
-    RecursiveEnumerate(PathBuf),
-    FileRead(PathBuf),
-    FileReadU32(PathBuf),
-    FileReadText(PathBuf),
-}
-
-enum FileSystemOpResult {
-    RecursiveEnumerate(DirectoryCache),
-    FileRead(Vec<u8>),
-    FileReadU32(Vec<u32>),
-    FileReadText(String),
-    Error(io::Error),
 }
 
 impl DirectoryFileTree {
@@ -186,89 +166,5 @@ impl<'a> FileTree<'a> for DirectoryFileTree {
                 _ => panic!("Incorrect file read action response received."),
             }
         })
-    }
-}
-
-pub struct DirectoryIterator<'a> {
-    sub_iter: hash_map::Keys<'a, std::ffi::OsString, DirectoryEntry>,
-}
-
-impl<'a> From<hash_map::Keys<'a, std::ffi::OsString, DirectoryEntry>> for DirectoryIterator<'a> {
-    fn from(sub_iter: hash_map::Keys<'a, OsString, DirectoryEntry>) -> Self {
-        DirectoryIterator { sub_iter }
-    }
-}
-
-impl<'a> Iterator for DirectoryIterator<'a> {
-    type Item = &'a Path;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.sub_iter.next().map(|v| Path::new(v))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.sub_iter.size_hint()
-    }
-}
-
-fn recursive_enumerate_directory_impl(root: &Path, relative: &Path) -> Result<DirectoryEntry, io::Error> {
-    let real_path = {
-        let mut p = root.to_path_buf();
-        p.push(relative);
-        p
-    };
-
-    if real_path.is_file() {
-        Ok(DirectoryEntry::File)
-    } else {
-        let mut map = HashMap::new();
-        for entry_result in real_path.read_dir()? {
-            let entry = entry_result?;
-            let file_name = entry.file_name();
-            let new_path = {
-                let mut p = relative.to_path_buf();
-                p.push(file_name);
-                p
-            };
-            map.insert(entry.file_name(), recursive_enumerate_directory_impl(root, &new_path)?);
-        }
-        Ok(DirectoryEntry::Directory(map))
-    }
-}
-
-fn recursive_enumerate_directory(root: &Path) -> Result<DirectoryCache, io::Error> {
-    let entry = recursive_enumerate_directory_impl(root, Path::new("/"))?;
-
-    Ok(DirectoryCache {
-        root: root.to_path_buf(),
-        entry,
-    })
-}
-
-fn file_system_reactor_core(op: FileSystemOp) -> FileSystemOpResult {
-    match op {
-        FileSystemOp::RecursiveEnumerate(path) => match recursive_enumerate_directory(&path) {
-            Ok(cache) => FileSystemOpResult::RecursiveEnumerate(cache),
-            Err(err) => FileSystemOpResult::Error(err),
-        },
-        FileSystemOp::FileRead(path) => match std::fs::read(path) {
-            Ok(result) => FileSystemOpResult::FileRead(result),
-            Err(err) => FileSystemOpResult::Error(err),
-        },
-        FileSystemOp::FileReadU32(path) => {
-            let file = std::fs::File::create(path);
-            let buffered = file.map(|f| io::BufReader::with_capacity(64 * 1024, f));
-            match buffered {
-                Ok(reader) => match fs::file::read_stream_u32(reader) {
-                    Ok(result) => FileSystemOpResult::FileReadU32(result),
-                    Err(err) => FileSystemOpResult::Error(err),
-                },
-                Err(err) => FileSystemOpResult::Error(err),
-            }
-        }
-        FileSystemOp::FileReadText(path) => match std::fs::read_to_string(path) {
-            Ok(result) => FileSystemOpResult::FileReadText(result),
-            Err(err) => FileSystemOpResult::Error(err),
-        },
     }
 }
